# lib/api-zod/src/generated/types/createStrategyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Indian Options Paper Trading API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateStrategyBodyFrequency } from "./createStrategyBodyFrequency";
import type { CreateStrategyBodyStrategyType } from "./createStrategyBodyStrategyType";
import type { CreateStrategyBodyUnderlying } from "./createStrategyBodyUnderlying";

export interface CreateStrategyBody {
  name: string;
  strategyType: CreateStrategyBodyStrategyType;
  underlying: CreateStrategyBodyUnderlying;
  frequency: CreateStrategyBodyFrequency;
  lotMultiplier: number;
  /** @nullable */
  deltaTarget?: number | null;
  /** @nullable */
  wingWidth?: number | null;
  /** @nullable */
  stopLossPct?: number | null;
  /** @nullable */
  targetProfitPct?: number | null;
  /** @nullable */
  entryTimeIst?: string | null;
  /** @nullable */
  exitTimeIst?: string | null;
}